fix(auth): allow superuser access to metrics without company association
- Create common get_user_company_id helper in app/api/v1/helpers.py
- Superusers can now access any company data without employee record
- Auto-create default company if none exists for superuser access
- Fixes 403 Forbidden errors for admin@example.com superuser access
- Remove duplicate get_user_company_id from metrics.py

Resolves: admin users getting 403 when accessing /api/v1/metrics/* endpoints
